# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SmartOrder AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/summary:
    get:
      operationId: getReportSummary
      tags: [reports]
      summary: Revenue, orders, avg order value summary
      security:
        - BearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSummary"
  /reports/top-products:
    get:
      operationId: getTopProducts
      tags: [reports]
      summary: Top selling products
      security:
        - BearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
  /reports/top-clients:
    get:
      operationId: getTopClients
      tags: [reports]
      summary: Top spending clients
      security:
        - BearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopClient"
  /reports/hourly:
    get:
      operationId: getHourlyReport
      tags: [reports]
      summary: Orders count grouped by hour and day of week
      security:
        - BearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HourlyDataPoint"
  /reports/daily-revenue:
    get:
      operationId: getDailyRevenue
      tags: [reports]
      summary: Daily revenue for the last N days
      security:
        - BearerAuth: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyRevenue"
  /reports/dashboard:
    get:
      operationId: getDashboardStats
      tags: [reports]
      summary: Dashboard KPI stats (today revenue, today orders, total clients, loyalty rate)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  # Staff
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderItem:
      type: object
      required: [productId, name, price, quantity]
      properties:
        productId:
          type: string
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer
    Order:
      type: object
      required: [id, businessId, clientName, clientPhone, items, total, depositAmount, status, orderType, createdAt]
      properties:
        id:
          type: string
        businessId:
          type: string
        clientName:
          type: string
        clientPhone:
          type: string
        clientEmail:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        depositAmount:
          type: number
        status:
          type: string
        orderType:
          type: string
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    ReportSummary:
      type: object
      required: [totalRevenue, ordersCount, avgOrderValue, period]
      properties:
        totalRevenue:
          type: number
        ordersCount:
          type: integer
        avgOrderValue:
          type: number
        period:
          type: string
    TopProduct:
      type: object
      required: [productId, name, totalSold, totalRevenue]
      properties:
        productId:
          type: string
        name:
          type: string
        totalSold:
          type: integer
        totalRevenue:
          type: number
        category:
          type: string
    TopClient:
      type: object
      required: [clientId, name, phone, totalOrders, totalSpent]
      properties:
        clientId:
          type: string
        name:
          type: string
        phone:
          type: string
        totalOrders:
          type: integer
        totalSpent:
          type: number
    HourlyDataPoint:
      type: object
      required: [hour, dayOfWeek, count]
      properties:
        hour:
          type: integer
        dayOfWeek:
          type: integer
        count:
          type: integer
    DailyRevenue:
      type: object
      required: [date, revenue, ordersCount]
      properties:
        date:
          type: string
        revenue:
          type: number
        ordersCount:
          type: integer
    DashboardStats:
      type: object
      required: [todayRevenue, todayOrders, totalClients, loyaltyRate, recentOrders]
      properties:
        todayRevenue:
          type: number
        todayOrders:
          type: integer
        totalClients:
          type: integer
        loyaltyRate:
          type: number
        recentOrders:
          type: array
          items:
            $ref: "#/components/schemas/Order"